fixes section header not re-rendering problem
Ok, I think I see what the issue was. I'm guessing that at first navigate was being used for the
section horizontal menu, but the headers were not being re-rendered. So to fix this, instead of
using navigate, they used pop and push, but unknowingly, this caused the small bug of not staying in place
when a new section in the horizontal menu was clicked since push and pop causes re-render but for the section header,
which is not needed.